+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- Persist response body query text box contents
+  - Previously it would reset whenever you made a new request or changed recipes
+
 ## [1.3.1] - 2024-05-21
 
 ### Fixed